# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Torito",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://localhost:8080/LaravelEGF/public"
        }
    ],
    "paths": {
        "/api/cajera": {
            "get": {
                "summary": "Mostrar Propiedades Ganaderas",
                "operationId": "App\\Http\\Controllers\\Api\\CajeraController::index",
                "responses": {
                    "200": {
                        "description": "Mostrar todos las propiedades."
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                }
            }
        },
        "/api/auth/login": {
            "post": {
                "tags": [
                    "auth"
                ],
                "summary": "autenticacion de usuario",
                "description": "Ingrese correo electronico y contraseña",
                "operationId": "App\\Http\\Controllers\\AuthController::login",
                "requestBody": {
                    "description": "Credenciales de usuario",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "123456789"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "422": {
                        "description": "Credenciales incorrectas",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Usuario y/o contraseña incorrectas."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autorizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Usuario no autorizado."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/auth/register": {
            "post": {
                "tags": [
                    "auth"
                ],
                "summary": "crea un nuevo usuario",
                "description": "Ingrese email y contraseña",
                "operationId": "App\\Http\\Controllers\\AuthController::register",
                "requestBody": {
                    "description": "Crea nuevo usuario",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password",
                                    "password_confirmation"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "123456789"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "123456789"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Se creo exitosamente el usuario."
                    },
                    "400": {
                        "description": "Credenciales no validas",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/auth/logout": {
            "get": {
                "tags": [
                    "auth"
                ],
                "summary": "Cierra session de usuario",
                "operationId": "App\\Http\\Controllers\\AuthController::logout",
                "responses": {
                    "200": {
                        "description": "Session cerrada correctamente."
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/auth/start": {
            "get": {
                "tags": [
                    "auth"
                ],
                "summary": "Usuario no autorizado",
                "operationId": "App\\Http\\Controllers\\AuthController::start",
                "responses": {
                    "401": {
                        "description": "No tiene autorizacion.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "No autorizado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/auth/list": {
            "get": {
                "tags": [
                    "auth"
                ],
                "summary": "Mostrar lista de usuarios",
                "responses": {
                    "200": {
                        "description": "Mostrar todos los usuarios."
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    },
                    "401": {
                        "description": "Respuesta si el usuario no esta autorizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "No autorizado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/auth/user-profile": {
            "get": {
                "tags": [
                    "auth"
                ],
                "summary": "Mostrar usuario autenticado",
                "operationId": "App\\Http\\Controllers\\AuthController::userProfile",
                "responses": {
                    "200": {
                        "description": "Mostrar todos los usuarios."
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "apiAuth": {
                "type": "http",
                "description": "Ingrese token de autenticacion",
                "name": "Token based Based",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}